requirements.txt
streamlit
langchain-google-genai
faiss-cpu
python-dotenv
langchain-community
langchain-core
langchain-classic
pypdf
